feat(rpg): Фаза 4 шаг 4г — Health-хранилище в int, сейв v80
ecs::Health = int{hp,maxHp}: все боевые писатели целые с Фазы 3, float
держал только память о дробных ранах. Дробный реген остаётся в carry-
аккумуляторах, не в баре. apply_lethal_damage = ровно остаток без ceil.
Свип читателей: дроби ран считаются float(hp)/float(maxHp) (int-деление =
тихий ноль — поймано в body_contract/subworld_spawn_parity/squad_war),
printf %f с int-аргументом вычищен (UB), экзекуция дробного бара в
damage_door_test переписана под целый бар. MacroNpcRecord.health едет тем
же POD-ом — kSaveVersion 80.

ctest 94/94; смоуки melee/cast_spell/missile/reputation/stats_settle по
сидам 1,7,999 PASS.

+++ src/ecs/components.h
@@ -20,8 +20,12 @@ struct Position { float x, y, z; };
 // Smoothed render position (for visual interpolation).
 struct VisualPos { float vx, vy, speed; };
 
-// Health component.
-struct Health { float hp, maxHp; };
+// Health component. INTEGER since phase 4г: every combat writer has been
+// whole since the dice phase (int amount through the one damage door), so the
+// float storage held nothing but the memory of fractional wounds that no
+// longer exist. Fractional REGEN lives in its accumulators
+// (macro/player_recovery.h carries), never in the bar.
+struct Health { int hp, maxHp; };
 
 // Explicit combat body radius — the distance at which this entity is struck by
 // melee, projectiles, and blasts (see the sub-layer target_radius()). It is the
